% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEqCg.R
\name{calcEqCg}
\alias{calcEqCg}
\title{The number of equivalent complete generations for several gids}
\usage{
calcEqCg(gid, dbenv)
}
\arguments{
\item{gid}{a vector of gids}

\item{dbenv}{the database connection environment}
}
\value{
a list of the EqCgs and the total number of cycles
}
\description{
The number of equivalent complete generations for several gids
}
