% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNei.R
\name{calcNei}
\alias{calcNei}
\title{calculate Nei}
\usage{
calcNei(eqcgvec, If1s)
}
\arguments{
\item{a}{vector of the equivalent complege generations}

\item{a}{vector of the inbreeding coefficients of the F1s}
}
\value{
a vector of Nei values
}
\description{
calculate Nei
}
