% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnGMS.R
\name{cnGMS}
\alias{cnGMS}
\title{Setting the connection}
\usage{
cnGMS(usr, psw)
}
\arguments{
\item{usr}{character username}

\item{pws}{character passward}
}
\value{
the object which is the function environment
}
\description{
Setting the connection
}
