% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iF1.R
\name{F1inbred}
\alias{F1inbred}
\title{F1inbred Function to identify the inbreedign coefficent of the F1 progenitor}
\usage{
F1inbred(cp, gidvec, dbenv)
}
\arguments{
\item{cp}{a coefficent of parentage matrix}

\item{gidvec}{A vector of numeric gids}

\item{dbenv}{the database connection environment}
}
\value{
a vector of inbreeding coefficents
}
\description{
F1inbred Function to identify the inbreedign coefficent of the F1 progenitor
}
