% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyPCA.R
\name{fancyPCA}
\alias{fancyPCA}
\title{fancyPCA function for plotting interactive PCA}
\usage{
fancyPCA(q, labs, dim = 3)
}
\arguments{
\item{q}{is an object that is returned from princomp}

\item{labs}{are the labels for the points in the plot}

\item{dim}{is the number of dimensions are the labels for the points}
}
\description{
fancyPCA function for plotting interactive PCA
}
