% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get1Gen.R
\name{get1Gen}
\alias{get1Gen}
\title{get1Gen Function for getting one generation of a pedigree}
\usage{
get1Gen(gidvec, dbenv)
}
\arguments{
\item{gidvec}{A vector of gids}

\item{dbenv}{the database connection environment}
}
\value{
A dataframe with the pedigree
}
\description{
get1Gen Function for getting one generation of a pedigree
}
