% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getallGen.R
\name{getallGen}
\alias{getallGen}
\title{getallGen function for getting all generations of a pedigree}
\usage{
getallGen(gidvec, dbenv)
}
\arguments{
\item{gidvec}{A vector of gids}

\item{dbenv}{the database connection environment}
}
\value{
A dataframe with the pedigree
}
\description{
getallGen function for getting all generations of a pedigree
}
