% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getallGids.R
\name{getallGids}
\alias{getallGids}
\title{get1Gen Function for getting all gids associated with a name}
\usage{
getallGids(nm, dbenv)
}
\arguments{
\item{nm}{A vector of names}

\item{dbenv}{the database connection environment}
}
\value{
a vector of gids
}
\description{
get1Gen Function for getting all gids associated with a name
}
