% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getallNames.R
\name{getallNames}
\alias{getallNames}
\title{get1Gen Function for getting all gids associated with a name}
\usage{
getallNames(gid, dbenv, type = 6)
}
\arguments{
\item{gid}{A gid}

\item{dbenv}{the database connection environment}

\item{type}{a numeric name type}
}
\value{
a vector of names
}
\description{
get1Gen Function for getting all gids associated with a name
}
