% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOP.R
\name{getCOP}
\alias{getCOP}
\title{getCOP Function for getting a sparse relationship (COP) matrix}
\usage{
getCOP(gidvec = NULL, ped)
}
\arguments{
\item{gidvec}{a vector of gids of interest}

\item{Ped}{Sparse pedigree matrix (class pedigree from pedigreemm) OR a data fame}
}
\value{
A sparse relationship matrix
}
\description{
getCOP Function for getting a sparse relationship (COP) matrix
}
