% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF1.R
\name{getF1}
\alias{getF1}
\title{getF1 function to get gid of the nearest f1 cross}
\usage{
getF1(gid, dbenv)
}
\arguments{
\item{gid}{a gid of interest}

\item{dbenv}{the database connection environment}
}
\value{
A sparse relationship matrix
}
\description{
getF1 function to get gid of the nearest f1 cross
}
