% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMgid.R
\name{getMgid}
\alias{getMgid}
\title{getMgid Function for getting the last gid resulting from a derivative method}
\usage{
getMgid(gid, dbenv)
}
\arguments{
\item{gid}{A numeric gid}

\item{dbenv}{the database connection environment}
}
\value{
a numeric gid
}
\description{
getMgid Function for getting the last gid resulting from a derivative method
}
