% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrnts.R
\name{getPrnts}
\alias{getPrnts}
\title{getPrnts Function to get parents of an individaul}
\usage{
getPrnts(gid, dbenv, pastenm = F)
}
\arguments{
\item{gid}{A numeric gid}

\item{dbenv}{the database connection environment}

\item{pastenm}{a boolean indicating if parent gids should be pasted together}
}
\value{
the parent gids
}
\description{
getPrnts Function to get parents of an individaul
}
