% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gidMax.R
\name{gidMax}
\alias{gidMax}
\title{gidMax Function to identify the most commonly used gid}
\usage{
gidMax(gidvec, dbenv)
}
\arguments{
\item{gidvec}{A vector of numeric gids}

\item{dbenv}{the database connection environment}
}
\description{
gidMax Function to identify the most commonly used gid
}
