% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inDatabase.R
\name{inDatabase}
\alias{inDatabase}
\title{inDatabase Function to check if something is in the database}
\usage{
inDatabase(gid, dbenv)
}
\arguments{
\item{gid}{A gid}

\item{dbenv}{environment with connection to datbase}
}
\value{
a boolean indicating if it is in the database or not
}
\description{
inDatabase Function to check if something is in the database
}
