% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinNe.R
\name{kinNe}
\alias{kinNe}
\title{kinNe Function for calculating Ne using a kinship matrix}
\usage{
kinNe(A, selIDs)
}
\arguments{
\item{A}{kinship matrix}

\item{selIDs}{vector containing selected gids}
}
\value{
a list of containing inbreeding coefficeint of generation 0 and 1, and the Ne
}
\description{
kinNe Function for calculating Ne using a kinship matrix
}
