% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode Function for finding the most frequent value in a vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the most frequent value
}
\description{
Mode Function for finding the most frequent value in a vector
}
