% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA function for making a two-paneled pca plot}
\usage{
plotPCA(q)
}
\arguments{
\item{q}{is a result of princomp}
}
\description{
plotPCA function for making a two-paneled pca plot
}
