% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSelect.R
\name{sampleSelect}
\alias{sampleSelect}
\title{Sample selection function for selecting representative samples of a target population}
\usage{
sampleSelect(cp, cand, target, nchoose, herit = 0.5, ntry = 300,
  nrep = 30)
}
\arguments{
\item{cp}{a relationship matrx}

\item{cand}{is a vector of indices pointing to the candidates for the sample}

\item{target}{is a vector of indices pointing to the target population}

\item{nchoose}{is the desired sample size}

\item{herit}{is the heritability of the trait}

\item{ntry}{is the number of iterations of the exchange algorithm}

\item{nrep}{is the number of repitions of the exchange algorithm}
}
\description{
Sample selection function for selecting representative samples of a target population
}
