% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePedg1.R
\name{tracePedg1}
\alias{tracePedg1}
\title{tracePedg1 function for tracing one generation of a pedigree}
\usage{
tracePedg1(gid, dbenv)
}
\arguments{
\item{gid}{A gid}

\item{dbenv}{the database connection environment}
}
\value{
A dataframe of query results containg complete generation traced
}
\description{
tracePedg1 function for tracing one generation of a pedigree
}
